moved mocks to separate file, added before hook
